do not multiply regularization penalty by nepochs
I realized that the contribution of regularization to the total chi squared
should not be multiplied by the number of epochs. With more observations,
you are adding information, which should eventually overwhelm the prior
imposed by the regularization penalty. This is obvious when you
consider the following two cases: In case A, you have two observations with
the same noise `N`. In case B, you have one observation with noise
`N / sqrt(2)`. The two cases contain the same information about the
true value(s), and should result in the same fitted value(s). This is only
true when the regularization penalty is constant, not multiplied by
the number of observations.